* Renamed Timer to Scheduler to avoid any ambiguity/confusion with java.util.Timer (plus, makes auto import much less annoying).  * When a player changes channels, their buffs are sent to the receiving channel so that buff expires are synced up and buff effects are recognized.  * Summons now are removed from the map when the owner player leaves, and spawn when the owner player enters into a new map.  * Fixed integer overflow causing LittleEndianReader.readLong to return incorrect values.  * Summons/puppets now removed from map when changing maps/channels, or when logging off. Summons are respawned on the new map if changing maps/channels.

+++ src/argonms/tools/input/LittleEndianReader.java
@@ -36,10 +36,10 @@ public abstract class LittleEndianReader {
 	public long readLong() {
 		byte[] bytes = read(8);
 		return (
-				((bytes[0] & 0xFF)) +
-				((bytes[1] & 0xFF) << 8) +
-				((bytes[2] & 0xFF) << 16) +
-				((bytes[3] & 0xFF) << 24) +
+				((long) (bytes[0] & 0xFF)) +
+				((long) (bytes[1] & 0xFF) << 8) +
+				((long) (bytes[2] & 0xFF) << 16) +
+				((long) (bytes[3] & 0xFF) << 24) +
 				((long) (bytes[4] & 0xFF) << 32) +
 				((long) (bytes[5] & 0xFF) << 40) +
 				((long) (bytes[6] & 0xFF) << 48) +
